+++ src/components/StopWatchTimeDisplay.tsx
@@ -1,11 +1,16 @@
 import React from 'react';
 import {SafeAreaView, Text, StyleSheet} from 'react-native';
-import {formatTime} from './utils/timeUtils';
+import {formatTime} from '../utils/timeUtils';
 
 interface StopWatchDisplayProps {
   timeInSeconds: number;
 }
 
+/**
+ * Displays the total number of seconds in a more user-friendly HH:MM:SS format
+ * making use of formatTime() from timeUtils. When time is -1, this is a 'cleared state' to adhere to the
+ * requirements of the testing file for a null state when the stop button is pressed.
+ * */
 export default function StopWatchDisplay(props: StopWatchDisplayProps) {
   const {timeInSeconds} = props;
   return (
