+++ src/utils/timeUtils.ts
@@ -1,4 +1,9 @@
-
+/**
+ * Formats a total number of seconds into a time string (HH:MM:SS).
+ *
+ * @param {number} totalSeconds - The total number of seconds.
+ * @returns {string} Formatted time string.
+ */
 export function formatTime(totalSeconds: number): string {
   const hours = Math.floor(totalSeconds / 3600);
   const minutes = Math.floor((totalSeconds % 3600) / 60);
